[백준] 15964_이상한 기호
이전에 부분점수를 받았던 문제에서 int형을 long으로 바꾸어 주었더니 만점을 받았다
-> 입력의 길이가 int로는 안담겼다

+++ BOJ/src/백준/p15964_이상한기호.java
@@ -13,10 +13,10 @@ public class p15964_이상한기호 {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		StringTokenizer st = new StringTokenizer(br.readLine());
 		
-		int first = Integer.parseInt(st.nextToken());
-		int second = Integer.parseInt(st.nextToken());
+		long first = Integer.parseInt(st.nextToken());
+		long second = Integer.parseInt(st.nextToken());
 		
-		int result =  (first + second)*(first-second);
+		long result =  (first + second)*(first-second);
 		
 		System.out.println(result);
 	}
